pyproject.toml
[tool.poetry]
name = "openwisp"
version = "0.1.0"
description = ""
authors = ["Thai Nguyen <shiroemon279@gmail.com"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
Flask = "^2.0.1"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
